[refactor #19] RedisConnectionFactory 생성 로직 제거, Spring Boot의 자동 주입을 활용하도록 변경
