+++ src/main/java/com/wellcome/WellcomeBE/domain/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.wellcome.WellcomeBE.domain.member;
+package com.wellcome.WellcomeBE.domain.member.repository;
 
+import com.wellcome.WellcomeBE.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
