+++ src/main/java/com/wellcome/WellcomeBE/global/security/TokenProvider.java
@@ -1,7 +1,7 @@
 package com.wellcome.WellcomeBE.global.security;
 
 import com.wellcome.WellcomeBE.domain.member.Member;
-import com.wellcome.WellcomeBE.domain.member.MemberRepository;
+import com.wellcome.WellcomeBE.domain.member.repository.MemberRepository;
 import com.wellcome.WellcomeBE.domain.member.dto.response.KakaoUserInfoResponse;
 import com.wellcome.WellcomeBE.global.exception.CustomException;
 import jakarta.servlet.http.HttpServletRequest;
